# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PP CRM",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/chars/{char}": {
            "get": {
                "tags": [
                    "Chars"
                ],
                "summary": "Единичная запись",
                "operationId": "188f7c950e5b256fe590e71564b7c076",
                "parameters": [
                    {
                        "name": "char",
                        "in": "path",
                        "description": "Id характеристики",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Получить характеристику",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Color"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Chars"
                ],
                "summary": "Удаление записи",
                "operationId": "4887c8e0b5cec579400c15835ae5b689",
                "parameters": [
                    {
                        "name": "char",
                        "in": "path",
                        "description": "Id характеристики",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Характеристика удалена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Char was deleted"
                                        },
                                        "err": {
                                            "type": "string",
                                            "example": "none"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Chars"
                ],
                "summary": "Обновление характеристики",
                "operationId": "35162d7ab1b79030f9ec47f9e9d8658f",
                "parameters": [
                    {
                        "name": "char",
                        "in": "path",
                        "description": "Id характеристики",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Color"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Характеристика обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Char was updated"
                                        },
                                        "err": {
                                            "type": "string",
                                            "example": "none"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/params/{param}": {
            "get": {
                "tags": [
                    "Params"
                ],
                "summary": "Единичная запись",
                "operationId": "f7f12f0cf906993f6322f9b80a846592",
                "parameters": [
                    {
                        "name": "param",
                        "in": "path",
                        "description": "Id параметра",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Получить параметр",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Width"
                                        },
                                        "measure": {
                                            "type": "string",
                                            "example": "mm"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Params"
                ],
                "summary": "Удаление записи",
                "operationId": "2b56dab7f394035bde220228a176e1b3",
                "parameters": [
                    {
                        "name": "param",
                        "in": "path",
                        "description": "Id параметра",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Параметр удален",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Param was deleted"
                                        },
                                        "err": {
                                            "type": "string",
                                            "example": "none"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Params"
                ],
                "summary": "Обновление параметра",
                "operationId": "c952dc4192d2872b79435babf3681f6b",
                "parameters": [
                    {
                        "name": "param",
                        "in": "path",
                        "description": "Id характеристики",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Width"
                                            },
                                            "measure": {
                                                "type": "string",
                                                "example": "mm"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Параметр обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Param was updated"
                                        },
                                        "err": {
                                            "type": "string",
                                            "example": "none"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}